Move getLocales in base class.
